+++ discord/message.py
@@ -147,9 +147,13 @@ class Attachment(Hashable):
         minutes or not valid at all.
     content_type: Optional[:class:`str`]
         The attachment's `media type <https://en.wikipedia.org/wiki/Media_type>`_
+    ephemeral: :class:`bool`
+        Whether or not the attachment is ephemeral.
+
+        .. versionadded:: 0.0.7
     """
 
-    __slots__ = ('id', 'size', 'height', 'width', 'filename', 'url', 'proxy_url', '_http', 'content_type')
+    __slots__ = ('id', 'size', 'height', 'width', 'filename', 'url', 'proxy_url', '_http', 'content_type', 'ephemeral')
 
     def __init__(self, *, data: AttachmentPayload, state: ConnectionState):
         self.id: int = int(data['id'])
@@ -159,6 +163,7 @@ class Attachment(Hashable):
         self.filename: str = data['filename']
         self.url: str = data.get('url')
         self.proxy_url: str = data.get('proxy_url')
+        self.ephemeral: bool = data.get('ephemeral', False)
         self._http = state.http
         self.content_type: Optional[str] = data.get('content_type')