+++ discord/types/interactions.py
@@ -49,7 +49,7 @@ class ApplicationCommand(TypedDict):
     type: Optional[ApplicationCommandType]
 
 
-ApplicationCommandOptionType = Literal[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+ApplicationCommandOptionType = Literal[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11]
 
 
 class ApplicationCommandOption(TypedDict):